tags: restructure utils into submodule of module
